+++ test/it/ParamsSuiteIT.ts
@@ -1,0 +1,20 @@
+import { params, suite } from "../../index";
+import { AbstractSuiteITBase, SuiteTestParams } from "./AbstractSuiteITBase";
+
+@suite(timeout(10000))
+class ParamsSuiteIT extends AbstractSuiteITBase {
+
+  @params({ target: "es5", fixture: "params.suite" })
+  @params({ target: "es6", fixture: "params.suite" })
+  @params({ target: "es5", fixture: "params.skip.suite" })
+  @params({ target: "es6", fixture: "params.skip.suite" })
+  @params({ target: "es5", fixture: "params.only.suite" })
+  @params({ target: "es6", fixture: "params.only.suite" })
+  @params({ target: "es5", fixture: "params.naming.suite" })
+  @params({ target: "es6", fixture: "params.naming.suite" })
+  @params.naming(({ target, fixture }: SuiteTestParams) => `${fixture} ${target}`)
+  runTest(params: SuiteTestParams) {
+
+    super.runTest(params);
+  }
+}