+++ CHANGELOG.md
@@ -18,3 +18,14 @@
 
 
 
+<a name="1.1.17"></a>
+## [1.1.17](https://github.com/PanayotCankov/mocha-typescript/compare/v1.1.16...v1.1.17) (2018-07-17)
+
+
+### Bug Fixes
+
+* Augment the new [@types](https://github.com/types)/mocha interfaces ([515b876](https://github.com/PanayotCankov/mocha-typescript/commit/515b876))
+* include di/typedi.js in the package ([a42a84b](https://github.com/PanayotCankov/mocha-typescript/commit/a42a84b))
+
+
+